requirements.txt
# FastAPI サーバー
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# HTTP クライアント (データ取得・アップロード)
aiohttp>=3.8.0,<4.0.0

# データ処理
pyyaml>=6.0